requirements.txt
google-cloud-firestore==2.0.2
google-cloud-secret-manager==2.2.0
google-auth==1.30.1
neo4j==4.2.1
elasticsearch==7.0.2
fastapi==0.68.1
jinja2==2.11.2
aiofiles==0.5.0
pandas==1.1.5
orjson==3.5.2
gunicorn==20.0.4
uvicorn==0.11.5
uvloop==0.14.0
httptools==0.1.1
